refactor(api/lib): Created a lib file, moved funtions away from api.lua
